refactor: Modify file structure
