+++ src/components/Share/posters/holderPoster.tsx
@@ -1,19 +1,18 @@
 import React, { useMemo, useRef } from 'react'
-import { HolderPosterData, PosterProps } from './poster.interface'
+import { HolderPosterData, PosterProps } from '../poster.interface'
 import {
   BackgroundImage,
   UserContainer,
   PosterContainer,
-  usePosterLoader,
-  useUrlToBase64,
-} from './shareUtils'
-import BackgroundImagePath from '../../assets/img/share-bg/share-holder@3x.png'
-import { getImagePreviewUrl } from '../../utils'
-import { Gallery } from './gallery'
+} from '../components/layout'
+import BackgroundImagePath from '../../../assets/img/share-bg/share-holder@3x.png'
+import { getImagePreviewUrl } from '../../../utils'
+import { Gallery } from '../components/gallery'
 import styled from 'styled-components'
 import { useTranslation } from 'react-i18next'
-import { ShareAvatar } from './avatar'
-import PeopleImage from '../../assets/img/people.png'
+import { ShareAvatar } from '../components/avatar'
+import PeopleImage from '../../../assets/img/people.png'
+import { useUrlToBase64, usePosterLoader } from '../hooks'
 
 const ContentContainer = styled.div`
   background-color: #fff;
